one program which outputs two xml files for one audio
the Pangloss structure of xml files isn't too versatile when it comes to utterance length. lists in a WORDLIST format are best indicated for words or frame sentences, and texts in a TEXT format are better for narratives made of natural-speech utterances/sentences. This file proposes to treat differently three types of intervals : (1) those that end with "_S" ("simple words", in a WORDLIST format), (2) those that end with "_PG" ("Guided" utterances, in a WORDLIST format) and (3) those that end with PC ("in Context", in a TEXT format, where tier 2 contains a transcription and a gloss).